refactor(control): split udp endpoint retirement
